requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.5
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.5
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
Deprecated==1.2.18
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.6.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kombu==5.5.4
langdetect==1.0.9
limits==5.5.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.4
numpy==2.2.6
openai==1.104.2
opencv-python==4.12.0.88
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.1
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
pyspellchecker==0.8.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pywin32-ctypes==0.2.3
redis==6.4.0
regex==2025.9.18
reportlab==4.4.4
requests==2.32.3
rsa==4.9.1
six==1.17.0
slowapi @ git+https://github.com/laurentS/slowapi.git@5e2b9c108f79471fc96a5152bb69774894660cd1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tiktoken==0.12.0
tomli==2.3.0
tqdm==4.67.1
twilio==9.6.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.14
wrapt==1.17.3
yarl==1.20.0
